Feature/indexconvert (#154)
* implement split

* create project file

* simplify refactoring

* indexation refactoring

* conversion progress

* display highlight

* do not touch paket.dependencies if nothing changed

* implement migrate command

* implement consistency check when indexing

* Bad Seq hunt

* implement projects anthology per repository

* restore projects

* fix print issue + fix missing content file issue

* better dispplay highlight

* implement projects consolidation on push

* consolidate anthology on pull

* update changes

* project reorg

* more tests

* unit tests for tags

* do not consolidate on push - only on pull

* code refactoring

* implement index --check

* fix unit tests

* versions history compatibility

* move from - to _ for tag

+++ src/FullBuild/Generators/MsBuild.fs
@@ -203,7 +203,7 @@ let private cleanupProject (xproj : XDocument) (project : Project) : XDocument =
             "PropertyGroup", hasNoChild
         ]
 
-    seekAndDestroy |> Seq.iter (fun (x, y) -> cproj.Descendants(NsMsBuild + x).Where(y).Remove())
+    seekAndDestroy |> List.iter (fun (x, y) -> cproj.Descendants(NsMsBuild + x).Where(y).Remove())
     cproj
 
 
@@ -233,9 +233,10 @@ let private convertProject (xproj : XDocument) (project : Project) =
         let prjFile = repoDir |> GetFile project.ProjectFile
         let prjDir = Path.GetDirectoryName (prjFile.FullName) |> DirectoryInfo
         let infoFile = prjDir |> GetFile fileName
-        let content = File.ReadAllLines (infoFile.FullName) |> List.ofSeq
-                                                            |> patchAssemblyVersion
-        File.WriteAllLines(infoFile.FullName, content)
+        if infoFile.Exists then
+            let content = File.ReadAllLines (infoFile.FullName) |> List.ofSeq
+                                                                |> patchAssemblyVersion
+            File.WriteAllLines(infoFile.FullName, content)
 
     let cproj = cleanupProject xproj project
